requirement.txt
--index-url https://pypi.douban.com/simple/
--trusted-host pypi.douban.com

asgiref==3.3.0
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-bulk-update==2.2.0
django-extensions==3.0.9
django-filter==2.4.0
djangorestframework==3.12.2
drf-extensions==0.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
packaging==20.4
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
uritemplate==3.0.1
urllib3==1.25.11
